refactor: searchBox, UserList 컴포넌트 분리
