+++ src/pages/users/UserList/components/SearchBox/style.module.css
@@ -1,0 +1,28 @@
+.searchBox {
+  width: 340px;
+  height: 48px;
+  padding: 0 var(--space-4);
+  background-color: white;
+  border: 1px solid var(--gray-100);
+  color: var(--gray-200);
+  border-radius: var(--radius-3);
+  font-size: var(--text-lg);
+  display: flex;
+  justify-content: start;
+  align-items: center;
+  gap: var(--space-2);
+}
+
+.searchIcon {
+  width: 24px;
+  height: 24px;
+}
+
+.searchInput {
+  border: none;
+  flex: 1;
+}
+
+.searchInput:focus {
+  outline: none;
+}
