+++ src/pages/users/UserList/components/UserList/style.module.css
@@ -1,0 +1,17 @@
+.noResult {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 80px 0;
+  text-align: center;
+  color: var(--gray-200);
+  font-size: var(--text-3xl);
+  font-style: italic;
+}
+
+.userList {
+  max-width: 1280px;
+  margin: 80px auto;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(296px, 1fr));
+  gap: 52px var(--space-7);
+}
